requirements.txt
pandas>=2.0.0
snowflake-connector-python>=3.0.0
python-dotenv>=1.0.0
streamlit>=1.30.0
snowflake-connector-python[pandas]>=3.0.0
numpy==1.24.3
plotly==5.17.0
prophet==1.1.4
snowflake-sqlalchemy==1.5.0
altair==5.1.2
Pillow==10.0.1
openpyxl==3.1.2 